Move InstanceInfo inner classes outside
