+++ gcloud-java-compute/src/main/java/com/google/gcloud/compute/InstanceInfo.java
@@ -22,9 +22,7 @@ import com.google.api.services.compute.model.Instance;
 import com.google.common.base.Function;
 import com.google.common.base.MoreObjects;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
 
 import org.joda.time.format.DateTimeFormatter;
 import org.joda.time.format.ISODateTimeFormat;
@@ -33,7 +31,6 @@ import java.io.Serializable;
 import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 
 /**
@@ -121,611 +118,6 @@ public class InstanceInfo implements Serializable {
      * from inside the guest. You can choose to restart the instance or delete it.
      */
     TERMINATED
-  }
-
-  /**
-   * A list of tags for a Google Compute Engine Instance; with associated fingerprint. Tags are used
-   * to identify valid sources or targets for network firewalls and are specified by the client
-   * during instance creation. Each tag within the list must comply with RFC1035.
-   *
-   * @see <a href="https://www.ietf.org/rfc/rfc1035.txt">RFC1035</a>
-   */
-  public static final class Tags implements Serializable {
-
-    static final Function<com.google.api.services.compute.model.Tags, Tags> FROM_PB_FUNCTION =
-        new Function<com.google.api.services.compute.model.Tags, Tags>() {
-          @Override
-          public Tags apply(com.google.api.services.compute.model.Tags pb) {
-            return Tags.fromPb(pb);
-          }
-        };
-    static final Function<Tags, com.google.api.services.compute.model.Tags> TO_PB_FUNCTION =
-        new Function<Tags, com.google.api.services.compute.model.Tags>() {
-          @Override
-          public com.google.api.services.compute.model.Tags apply(Tags tags) {
-            return tags.toPb();
-          }
-        };
-
-    private static final long serialVersionUID = 5627093820497225322L;
-
-    private final List<String> values;
-    private final String fingerprint;
-
-    /**
-     * A builder for {@code Tags} objects.
-     */
-    public static final class Builder {
-
-      private List<String> values;
-      private String fingerprint;
-
-      private Builder() {
-        values = Lists.newArrayList();
-      }
-
-      private Builder(Tags tags) {
-        this.values = tags.values != null ? Lists.newArrayList(tags.values)
-            : Lists.<String>newArrayList();
-        this.fingerprint = tags.fingerprint;
-      }
-
-      /**
-       * Sets a list of tags to apply to an instance. Tags are used to identify valid sources or
-       * targets for network firewalls. Each tag within the list must comply with RFC1035.
-       *
-       * @see <a href="https://www.ietf.org/rfc/rfc1035.txt">RFC1035</a>
-       */
-      public Builder values(Iterable<String> values) {
-        this.values = Lists.newArrayList(values);
-        return this;
-      }
-
-      /**
-       * Sets a list of tags to apply to an instance. Tags are used to identify valid sources or
-       * targets for network firewalls. Each tag within the list must comply with RFC1035.
-       *
-       * @see <a href="https://www.ietf.org/rfc/rfc1035.txt">RFC1035</a>
-       */
-      public Builder values(String... values) {
-        this.values = Lists.newArrayList(Arrays.asList(checkNotNull(values)));
-        return this;
-      }
-
-      /**
-       * Adds a tag to the list of tags. Tags are used to identify valid sources or targets for
-       * network firewalls. The tag must comply with RFC1035.
-       *
-       * @see <a href="https://www.ietf.org/rfc/rfc1035.txt">RFC1035</a>
-       */
-      public Builder add(String tag) {
-        this.values.add(tag);
-        return this;
-      }
-
-      /**
-       * Sets the fingerprint for the tags. This value is needed to update instance's tags.
-       */
-      public Builder fingerprint(String fingerprint) {
-        this.fingerprint = fingerprint;
-        return this;
-      }
-
-      /**
-       * Creates a {@code Tags} object.
-       */
-      public Tags build() {
-        return new Tags(this);
-      }
-    }
-
-    private Tags(Builder builder) {
-      this.values = ImmutableList.copyOf(builder.values);
-      this.fingerprint = builder.fingerprint;
-    }
-
-    /**
-     * Returns a list of tags to apply to an instance. Tags are used to identify valid sources or
-     * targets for network firewalls. Each tag within the list must comply with RFC1035.
-     *
-     * @see <a href="https://www.ietf.org/rfc/rfc1035.txt">RFC1035</a>
-     */
-    public List<String> values() {
-      return values;
-    }
-
-    /**
-     * Returns the fingerprint for the tags. This value is needed to update instance's tags.
-     */
-    public String fingerprint() {
-      return fingerprint;
-    }
-
-    /**
-     * Returns a builder for the current instance tags.
-     */
-    public Builder toBuilder() {
-      return new Builder(this);
-    }
-
-    @Override
-    public String toString() {
-      return MoreObjects.toStringHelper(this)
-          .add("values", values)
-          .add("fingerprint", fingerprint)
-          .toString();
-    }
-
-    @Override
-    public int hashCode() {
-      return Objects.hash(values, fingerprint);
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-      return obj == this || obj instanceof Tags && Objects.equals(toPb(), ((Tags) obj).toPb());
-    }
-
-    com.google.api.services.compute.model.Tags toPb() {
-      com.google.api.services.compute.model.Tags tagsPb =
-          new com.google.api.services.compute.model.Tags();
-      tagsPb.setFingerprint(fingerprint);
-      tagsPb.setItems(values);
-      return tagsPb;
-    }
-
-    /**
-     * Returns a builder for a {@code Tags} object given the tags to apply to the instance. Each tag
-     * within the list must comply with RFC1035.
-     *
-     * @see <a href="https://www.ietf.org/rfc/rfc1035.txt">RFC1035</a>
-     */
-    public static Builder builder(Iterable<String> values) {
-      return new Builder().values(values);
-    }
-
-    /**
-     * Returns a builder for a {@code Tags} object given the tags to apply to the instance. Each tag
-     * within the list must comply with RFC1035.
-     *
-     * @see <a href="https://www.ietf.org/rfc/rfc1035.txt">RFC1035</a>
-     */
-    public static Builder builder(String... values) {
-      return new Builder().values(values);
-    }
-
-    /**
-     * Returns a {@code Tags} object given the tags to apply to the instance. Each tag within the
-     * list must comply with RFC1035.
-     *
-     * @see <a href="https://www.ietf.org/rfc/rfc1035.txt">RFC1035</a>
-     */
-    public static Tags of(Iterable<String> values) {
-      return builder(values).build();
-    }
-
-    /**
-     * Returns a {@code Tags} object given the tags to apply to the instance. Each tag within the
-     * list must comply with RFC1035.
-     *
-     * @see <a href="https://www.ietf.org/rfc/rfc1035.txt">RFC1035</a>
-     */
-    public static Tags of(String... values) {
-      return builder(values).build();
-    }
-
-    static Tags fromPb(com.google.api.services.compute.model.Tags tagsPb) {
-      Builder builder =
-          builder(tagsPb.getItems() != null ? tagsPb.getItems() : ImmutableList.<String>of());
-      return builder.fingerprint(tagsPb.getFingerprint()).build();
-    }
-  }
-
-  /**
-   * Metadata for Google Compute Engine Instance as ket/value pairs. This includes custom metadata
-   * and predefined keys.
-   *
-   * @see <a href="https://cloud.google.com/compute/docs/metadata">Metadata</a>
-   */
-  public static final class Metadata implements Serializable {
-
-    static final Function<com.google.api.services.compute.model.Metadata, Metadata>
-        FROM_PB_FUNCTION =
-        new Function<com.google.api.services.compute.model.Metadata, Metadata>() {
-          @Override
-          public Metadata apply(com.google.api.services.compute.model.Metadata pb) {
-            return Metadata.fromPb(pb);
-          }
-        };
-    static final Function<Metadata, com.google.api.services.compute.model.Metadata> TO_PB_FUNCTION =
-        new Function<Metadata, com.google.api.services.compute.model.Metadata>() {
-          @Override
-          public com.google.api.services.compute.model.Metadata apply(Metadata metadata) {
-            return metadata.toPb();
-          }
-        };
-
-    private static final long serialVersionUID = -945038809838910107L;
-
-    private final Map<String, String> values;
-    private final String fingerprint;
-
-    /**
-     * A builder for {@code Metadata} objects.
-     */
-    public static final class Builder {
-
-      private Map<String, String> values;
-      private String fingerprint;
-
-      Builder() {
-        values = Maps.newHashMap();
-      }
-
-      Builder(Metadata metadata) {
-        this.values = metadata.values != null ? Maps.newHashMap(metadata.values)
-            : Maps.<String, String>newHashMap();
-        this.fingerprint = metadata.fingerprint;
-      }
-
-      /**
-       * Sets the metadata for the instance as key/value pairs. The total size of all keys and
-       * values must be less than 512 KB. Keys must conform to the following regexp:
-       * {@code [a-zA-Z0-9-_]+}, and be less than 128 bytes in length. This is reflected as part of
-       * a URL in the metadata server. Additionally, to avoid ambiguity, keys must not conflict with
-       * any other metadata keys for the project. Values must be less than or equal to 32768 bytes.
-       */
-      public Builder values(Map<String, String> values) {
-        this.values = Maps.newHashMap(checkNotNull(values));
-        return this;
-      }
-
-      /**
-       * Adds a key/value pair to the instance metadata. The total size of all keys and values must
-       * be less than 512 KB. Keys must conform to the following regexp: {@code [a-zA-Z0-9-_]+}, and
-       * be less than 128 bytes in length. This is reflected as part of a URL in the metadata
-       * server. Additionally, to avoid ambiguity, keys must not conflict with any other metadata
-       * keys for the project. Values must be less than or equal to 32768 bytes.
-       */
-      public Builder add(String key, String value) {
-        this.values.put(key, value);
-        return this;
-      }
-
-      /**
-       * Sets the fingerprint for the metadata. This value can be used to update instance's
-       * metadata.
-       */
-      public Builder fingerprint(String fingerprint) {
-        this.fingerprint = fingerprint;
-        return this;
-      }
-
-      /**
-       * Creates a {@code Metadata} object.
-       */
-      public Metadata build() {
-        return new Metadata(this);
-      }
-    }
-
-    private Metadata(Builder builder) {
-      this.values = ImmutableMap.copyOf(builder.values);
-      this.fingerprint = builder.fingerprint;
-    }
-
-    /**
-     * Returns instance's metadata as key/value pairs.
-     */
-    public Map<String, String> values() {
-      return values;
-    }
-
-    /**
-     * Returns the fingerprint for the metadata. This value can be used to update instance's
-     * metadata.
-     */
-    public String fingerprint() {
-      return fingerprint;
-    }
-
-    /**
-     * Returns a builder for the current instance metadata.
-     */
-    public Builder toBuilder() {
-      return new Builder(this);
-    }
-
-    @Override
-    public String toString() {
-      return MoreObjects.toStringHelper(this)
-          .add("values", values)
-          .add("fingerprint", fingerprint)
-          .toString();
-    }
-
-    @Override
-    public int hashCode() {
-      return Objects.hash(values, fingerprint);
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-      return obj == this
-          || obj instanceof Metadata
-          && Objects.equals(toPb(), ((Metadata) obj).toPb());
-    }
-
-    com.google.api.services.compute.model.Metadata toPb() {
-      com.google.api.services.compute.model.Metadata metadataPb =
-          new com.google.api.services.compute.model.Metadata();
-      metadataPb.setFingerprint(fingerprint);
-      List<com.google.api.services.compute.model.Metadata.Items> itemsPb =
-          Lists.newArrayListWithCapacity(values.size());
-      for (Map.Entry<String, String> entry : values.entrySet()) {
-        itemsPb.add(new com.google.api.services.compute.model.Metadata.Items()
-            .setKey(entry.getKey()).setValue(entry.getValue()));
-      }
-      metadataPb.setItems(itemsPb);
-      metadataPb.setFingerprint(fingerprint);
-      return metadataPb;
-    }
-
-    /**
-     * Returns a builder for a {@code Metadata} object.
-     */
-    public static Builder builder() {
-      return new Builder();
-    }
-
-    /**
-     * Returns a {@code Metadata} object given the the metadata as a map. The total size of all keys
-     * and values must be less than 512 KB. Keys must conform to the following regexp:
-     * {@code [a-zA-Z0-9-_]+}, and be less than 128 bytes in length. This is reflected as part of a
-     * URL in the metadata server. Additionally, to avoid ambiguity, keys must not conflict with any
-     * other metadata keys for the project. Values must be less than or equal to 32768 bytes.
-     */
-    public static Metadata of(Map<String, String> values) {
-      return builder().values(values).build();
-    }
-
-    static Metadata fromPb(com.google.api.services.compute.model.Metadata metadataPb) {
-      Builder builder = builder();
-      if (metadataPb.getItems() != null) {
-        Map<String, String> metadataValues =
-            Maps.newHashMapWithExpectedSize(metadataPb.getItems().size());
-        for (com.google.api.services.compute.model.Metadata.Items item : metadataPb.getItems()) {
-          metadataValues.put(item.getKey(), item.getValue());
-        }
-        builder.values(metadataValues);
-      }
-      return builder.fingerprint(metadataPb.getFingerprint()).build();
-    }
-  }
-
-  /**
-   * A service account, with its specified scopes, authorized for this instance.
-   *
-   * @see <a href="https://cloud.google.com/compute/docs/authentication">Authenticating from Google
-   *     Compute Engine</a>
-   */
-  public static final class ServiceAccount implements Serializable {
-
-    static final Function<com.google.api.services.compute.model.ServiceAccount, ServiceAccount>
-        FROM_PB_FUNCTION =
-        new Function<com.google.api.services.compute.model.ServiceAccount, ServiceAccount>() {
-          @Override
-          public ServiceAccount apply(com.google.api.services.compute.model.ServiceAccount pb) {
-            return ServiceAccount.fromPb(pb);
-          }
-        };
-    static final Function<ServiceAccount, com.google.api.services.compute.model.ServiceAccount>
-        TO_PB_FUNCTION =
-        new Function<ServiceAccount, com.google.api.services.compute.model.ServiceAccount>() {
-          @Override
-          public com.google.api.services.compute.model.ServiceAccount apply(
-              ServiceAccount metadata) {
-            return metadata.toPb();
-          }
-        };
-
-    private static final long serialVersionUID = 4199610694227857331L;
-
-    private final String email;
-    private final List<String> scopes;
-
-    private ServiceAccount(String email, List<String> scopes) {
-      this.email = email;
-      this.scopes = ImmutableList.copyOf(scopes);
-    }
-
-    /**
-     * Returns the email address of the service account.
-     */
-    public String email() {
-      return email;
-    }
-
-    /**
-     * Returns the list of scopes to be made available for this service account.
-     */
-    public List<String> scopes() {
-      return scopes;
-    }
-
-    @Override
-    public String toString() {
-      return MoreObjects.toStringHelper(this)
-          .add("email", email)
-          .add("scopes", scopes)
-          .toString();
-    }
-
-    @Override
-    public int hashCode() {
-      return Objects.hash(email, scopes);
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-      return obj == this
-          || obj instanceof ServiceAccount
-          && Objects.equals(toPb(), ((ServiceAccount) obj).toPb());
-    }
-
-    com.google.api.services.compute.model.ServiceAccount toPb() {
-      com.google.api.services.compute.model.ServiceAccount serviceAccountPb =
-          new com.google.api.services.compute.model.ServiceAccount();
-      serviceAccountPb.setEmail(email);
-      serviceAccountPb.setScopes(scopes);
-      return serviceAccountPb;
-    }
-
-    /**
-     * Returns a {@code ServiceAccount} object for the provided email and scopes.
-     */
-    public static ServiceAccount of(String email, List<String> scopes) {
-      return new ServiceAccount(email, scopes);
-    }
-
-    /**
-     * Returns a {@code ServiceAccount} object for the provided email and scopes.
-     */
-    public static ServiceAccount of(String email, String... scopes) {
-      return of(email, Arrays.asList(scopes));
-    }
-
-    static ServiceAccount fromPb(com.google.api.services.compute.model.ServiceAccount accountPb) {
-      return new ServiceAccount(accountPb.getEmail(), accountPb.getScopes());
-    }
-  }
-
-  /**
-   * A Google Compute Engine instance scheduling options. When there are system events that might
-   * cause your instances to be disrupted, Google Compute Engine automatically manages the
-   * scheduling decisions for your instances. Use {@code SchedulingOptions.preemptible()} to create
-   * a preemptible instance. Use {@code SchedulingOptions.standard()} to configure scheduling
-   * options for a standard instance.
-   *
-   * @see <a href="https://cloud.google.com/compute/docs/instances/setting-instance-scheduling-options">
-   *     Setting Instance Scheduling Options</a>
-   */
-  public static final class SchedulingOptions implements Serializable {
-
-    private static final long serialVersionUID = 4199610694227857331L;
-
-    private final boolean automaticRestart;
-    private final Maintenance maintenance;
-    private final boolean isPreemptible;
-
-    /**
-     * Defines the maintenance behavior for this instance.
-     */
-    public enum Maintenance {
-      /**
-       * The default behavior for standard instances.
-       */
-      MIGRATE,
-
-      /**
-       * The default and only possible behavior for preemptible instances.
-       */
-      TERMINATE
-    }
-
-    private SchedulingOptions(Boolean automaticRestart, Maintenance maintenance,
-        Boolean isPreemptible) {
-      this.automaticRestart = automaticRestart;
-      this.maintenance = maintenance;
-      this.isPreemptible = isPreemptible;
-    }
-
-    /**
-     * Returns whether the instance should be automatically restarted if it is terminated by Compute
-     * Engine (not terminated by a user).
-     */
-    public Boolean automaticRestart() {
-      return automaticRestart;
-    }
-
-    /**
-     * Returns the maintenance behavior for the instance.
-     */
-    public Maintenance maintenance() {
-      return maintenance;
-    }
-
-    /**
-     * Returns {@code true} if the instance is preemptible, {@code false} otherwhise.
-     *
-     * @see <a href="https://cloud.google.com/compute/docs/instances/preemptible">Preemptible
-     *     Instance</a>
-     */
-    public boolean isPreemptible() {
-      return isPreemptible;
-    }
-
-    @Override
-    public String toString() {
-      return MoreObjects.toStringHelper(this)
-          .add("automaticRestart", automaticRestart)
-          .add("maintenance", maintenance)
-          .add("isPreemptible", isPreemptible)
-          .toString();
-    }
-
-    @Override
-    public int hashCode() {
-      return Objects.hash(automaticRestart, maintenance, isPreemptible);
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-      return obj == this
-          || obj instanceof SchedulingOptions
-          && Objects.equals(toPb(), ((SchedulingOptions) obj).toPb());
-    }
-
-    com.google.api.services.compute.model.Scheduling toPb() {
-      com.google.api.services.compute.model.Scheduling schedulingPb =
-          new com.google.api.services.compute.model.Scheduling();
-      schedulingPb.setAutomaticRestart(automaticRestart);
-      schedulingPb.setPreemptible(isPreemptible);
-      if (maintenance != null) {
-        schedulingPb.setOnHostMaintenance(maintenance.name());
-      }
-      return schedulingPb;
-    }
-
-    /**
-     * Returns a {@code SchedulingOptions} object for a preemptible instance.
-     *
-     * @see <a href="https://cloud.google.com/compute/docs/instances/preemptible">Preemptible
-     *     Instance</a>
-     */
-    public static SchedulingOptions preemptible() {
-      return new SchedulingOptions(false, Maintenance.TERMINATE, true);
-    }
-
-    /**
-     * Returns a {@code SchedulingOptions} object for a standard instance.
-     *
-     * @param automaticRestart specifies whether the instance should be automatically restarted if
-     *     it is terminated by Compute Engine (not terminated by a user)
-     * @param maintenance defines the maintenance behavior for the instance
-     */
-    public static SchedulingOptions standard(boolean automaticRestart, Maintenance maintenance) {
-      return new SchedulingOptions(automaticRestart, maintenance, false);
-    }
-
-    static SchedulingOptions fromPb(com.google.api.services.compute.model.Scheduling schedPb) {
-      Maintenance maintenance = null;
-      if (schedPb.getOnHostMaintenance() != null) {
-        maintenance = Maintenance.valueOf(schedPb.getOnHostMaintenance());
-      }
-      return new SchedulingOptions(schedPb.getAutomaticRestart(), maintenance,
-          schedPb.getPreemptible());
-    }
   }
 
   /**
