+++ gcloud-java-compute/src/test/java/com/google/gcloud/compute/InstanceInfoTest.java
@@ -17,23 +17,14 @@
 package com.google.gcloud.compute;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
 
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
 import com.google.gcloud.compute.AttachedDisk.PersistentDiskConfiguration;
-import com.google.gcloud.compute.InstanceInfo.Metadata;
-import com.google.gcloud.compute.InstanceInfo.SchedulingOptions;
-import com.google.gcloud.compute.InstanceInfo.SchedulingOptions.Maintenance;
-import com.google.gcloud.compute.InstanceInfo.ServiceAccount;
-import com.google.gcloud.compute.InstanceInfo.Tags;
 
 import org.junit.Test;
 
 import java.util.List;
-import java.util.Map;
 
 public class InstanceInfoTest {
 
@@ -61,8 +52,7 @@ public class InstanceInfoTest {
   private static final ServiceAccount SERVICE_ACCOUNT =
       ServiceAccount.of("email", ImmutableList.of("scope1"));
   private static final List<ServiceAccount> SERVICE_ACCOUNTS = ImmutableList.of(SERVICE_ACCOUNT);
-  private static final SchedulingOptions SCHEDULING_OPTIONS =
-      SchedulingOptions.preemptible();
+  private static final SchedulingOptions SCHEDULING_OPTIONS = SchedulingOptions.preemptible();
   private static final String CPU_PLATFORM = "cpuPlatform";
   private static final InstanceInfo INSTANCE_INFO = InstanceInfo.builder(INSTANCE_ID, MACHINE_TYPE)
       .id(ID)
@@ -81,20 +71,6 @@ public class InstanceInfoTest {
       .build();
 
   @Test
-  public void testMetadataToBuilder() {
-    Metadata metadata = METADATA.toBuilder().fingerprint("newFingerprint").build();
-    assertEquals("newFingerprint", metadata.fingerprint());
-    comparevalues(METADATA, metadata.toBuilder().fingerprint(null).build());
-  }
-
-  @Test
-  public void testTagsToBuilder() {
-    Tags tags = TAGS.toBuilder().values("tag1").build();
-    assertEquals(ImmutableList.of("tag1"), tags.values());
-    compareTags(TAGS, tags.toBuilder().values("tag1", "tag2").build());
-  }
-
-  @Test
   public void testToBuilder() {
     compareInstanceInfo(INSTANCE_INFO, INSTANCE_INFO.toBuilder().build());
     InstanceInfo instance = INSTANCE_INFO.toBuilder().description("newDescription").build();
@@ -108,36 +84,6 @@ public class InstanceInfoTest {
     InstanceInfo instanceInfo = InstanceInfo.of(INSTANCE_ID, MACHINE_TYPE, ATTACHED_DISK,
         NETWORK_INTERFACE);
     assertEquals(instanceInfo, instanceInfo.toBuilder().build());
-  }
-
-  @Test
-  public void testMetadataBuilder() {
-    assertEquals(ImmutableMap.of("key1", "value1", "key2", "value2"), METADATA.values());
-    assertNull(METADATA.fingerprint());
-    Metadata metadata = Metadata.builder()
-        .values(ImmutableMap.of("key1", "value1", "key2", "value2"))
-        .build();
-    assertEquals(ImmutableMap.of("key1", "value1", "key2", "value2"), metadata.values());
-    assertNull(metadata.fingerprint());
-    metadata = Metadata.builder()
-        .values(ImmutableMap.of("key1", "value1", "key2", "value2"))
-        .fingerprint("fingerprint")
-        .build();
-    assertEquals(ImmutableMap.of("key1", "value1", "key2", "value2"), metadata.values());
-    assertEquals("fingerprint", metadata.fingerprint());
-  }
-
-  @Test
-  public void testTagsBuilder() {
-    Tags tags = Tags.builder().values(ImmutableList.of("tag1", "tag2")).build();
-    assertEquals(ImmutableList.of("tag1", "tag2"), tags.values());
-    assertNull(tags.fingerprint());
-    tags = Tags.builder().add("tag1").add("tag2").build();
-    assertEquals(ImmutableList.of("tag1", "tag2"), tags.values());
-    assertNull(tags.fingerprint());
-    tags = Tags.builder().add("tag1").add("tag2").fingerprint("fingerprint").build();
-    assertEquals(ImmutableList.of("tag1", "tag2"), tags.values());
-    assertEquals("fingerprint", tags.fingerprint());
   }
 
   @Test
@@ -176,35 +122,6 @@ public class InstanceInfoTest {
   }
 
   @Test
-  public void testMetadataOf() {
-    Map<String, String> map = ImmutableMap.of("key1", "value1", "key2", "value2");
-    comparevalues(METADATA, Metadata.of(map));
-  }
-
-  @Test
-  public void testTagsOf() {
-    compareTags(TAGS, Tags.of("tag1", "tag2"));
-    compareTags(TAGS, Tags.of(ImmutableList.of("tag1", "tag2")));
-  }
-
-  @Test
-  public void testServiceAccountOf() {
-    compareServiceAccount(SERVICE_ACCOUNT, ServiceAccount.of("email", ImmutableList.of("scope1")));
-    compareServiceAccount(SERVICE_ACCOUNT, ServiceAccount.of("email", "scope1"));
-  }
-
-  @Test
-  public void testSchedulingOptionsFactoryMethods() {
-    assertTrue(SCHEDULING_OPTIONS.isPreemptible());
-    assertFalse(SCHEDULING_OPTIONS.automaticRestart());
-    assertEquals(Maintenance.TERMINATE, SCHEDULING_OPTIONS.maintenance());
-    SchedulingOptions schedulingOptions = SchedulingOptions.standard(true, Maintenance.MIGRATE);
-    assertFalse(schedulingOptions.isPreemptible());
-    assertTrue(schedulingOptions.automaticRestart());
-    assertEquals(Maintenance.MIGRATE, schedulingOptions.maintenance());
-  }
-
-  @Test
   public void testOf() {
     InstanceInfo instance =
         InstanceInfo.of(INSTANCE_ID, MACHINE_TYPE, ATTACHED_DISK, NETWORK_INTERFACE);
@@ -226,36 +143,6 @@ public class InstanceInfoTest {
   }
 
   @Test
-  public void testMetadataToAndFromPb() {
-    comparevalues(METADATA, Metadata.fromPb(METADATA.toPb()));
-    Metadata metadata = Metadata.builder()
-        .values(ImmutableMap.of("key1", "value1", "key2", "value2"))
-        .fingerprint("fingerprint")
-        .build();
-    comparevalues(metadata, Metadata.fromPb(metadata.toPb()));
-  }
-
-  @Test
-  public void testTagsToAndFromPb() {
-    compareTags(TAGS, Tags.fromPb(TAGS.toPb()));
-    Tags tags = Tags.builder().add("tag1").add("tag2").fingerprint("fingerprint").build();
-    compareTags(tags, Tags.fromPb(tags.toPb()));
-  }
-
-  @Test
-  public void testServiceAccountToAndFromPb() {
-    compareServiceAccount(SERVICE_ACCOUNT, ServiceAccount.fromPb(SERVICE_ACCOUNT.toPb()));
-  }
-
-  @Test
-  public void testSchedulingOptionsToAndFromPb() {
-    compareSchedulingOptions(SCHEDULING_OPTIONS,
-        SchedulingOptions.fromPb(SCHEDULING_OPTIONS.toPb()));
-    SchedulingOptions schedulingOptions = SchedulingOptions.standard(true, Maintenance.MIGRATE);
-    compareSchedulingOptions(schedulingOptions, SchedulingOptions.fromPb(schedulingOptions.toPb()));
-  }
-
-  @Test
   public void testToAndFromPb() {
     compareInstanceInfo(INSTANCE_INFO, InstanceInfo.fromPb(INSTANCE_INFO.toPb()));
     InstanceInfo instance =
@@ -273,35 +160,6 @@ public class InstanceInfoTest {
     InstanceInfo instanceWithProject =
         InstanceInfo.of(INSTANCE_ID, MACHINE_TYPE, ATTACHED_DISK, NETWORK_INTERFACE);
     compareInstanceInfo(instanceWithProject, instance.setProjectId("project"));
-  }
-
-  public void comparevalues(Metadata expected, Metadata value) {
-    assertEquals(expected, value);
-    assertEquals(expected.fingerprint(), value.fingerprint());
-    assertEquals(expected.values(), value.values());
-    assertEquals(expected.hashCode(), value.hashCode());
-  }
-
-  public void compareTags(Tags expected, Tags value) {
-    assertEquals(expected, value);
-    assertEquals(expected.fingerprint(), value.fingerprint());
-    assertEquals(expected.values(), value.values());
-    assertEquals(expected.hashCode(), value.hashCode());
-  }
-
-  public void compareServiceAccount(ServiceAccount expected, ServiceAccount value) {
-    assertEquals(expected, value);
-    assertEquals(expected.email(), value.email());
-    assertEquals(expected.scopes(), value.scopes());
-    assertEquals(expected.hashCode(), value.hashCode());
-  }
-
-  public void compareSchedulingOptions(SchedulingOptions expected, SchedulingOptions value) {
-    assertEquals(expected, value);
-    assertEquals(expected.isPreemptible(), value.isPreemptible());
-    assertEquals(expected.maintenance(), value.maintenance());
-    assertEquals(expected.automaticRestart(), value.automaticRestart());
-    assertEquals(expected.hashCode(), value.hashCode());
   }
 
   public void compareInstanceInfo(InstanceInfo expected, InstanceInfo value) {
